* commit '825ca0c3fe62f0ddefc27917b11e97f3f26cae7c':
  Cleaning 2D and 3D matpolys tests
  Refactor and fix R2D and R3D split tests
  Last cleanup round
  Fix MOF, VOF and SLIC tests failures
  Update loops and get rid of unsigned counters
  Fix old-style type cast to static or implicit cast
  Fix missing variable on debug mode
  Add wonton fixes
  Fix uninitialized variables and spacing
  Fix out-of-order initialization
  Remove unused local variables
  Fix sign comparison warnings
  Fix 'matpoly' constructor and destructor
  Fix no return in assignment operator
  Fix wonton warnings
  Ignore Clion files

+++ tangram/utility/rpgtools/primitives.h
@@ -186,7 +186,7 @@ Tangram::MatPoly<3> prism(const std::vector<Tangram::Point3>& base_pts,
                           double dst_tol,
                           const Tangram::Vector3& base_normal =
                             Tangram::Vector3(0.0, 0.0, 0.0)) {
-  int nbase_vrts = (int) base_pts.size();
+  int nbase_vrts = base_pts.size();
   std::vector< std::vector<int> > ifaces_vrts(nbase_vrts + 2);
   ifaces_vrts[0].resize(nbase_vrts);
   std::iota(ifaces_vrts[0].begin(), ifaces_vrts[0].end(), 0);
@@ -260,7 +260,7 @@ Tangram::MatPoly<3> skewed_prism(const std::vector<Tangram::Point3>& base_pts,
                                  double dst_tol,
                                  const Tangram::Vector3& base_normal =
                                     Tangram::Vector3(0.0, 0.0, 0.0)) {
-  int nbase_vrts = (int) base_pts.size();
+  int nbase_vrts = base_pts.size();
   std::vector< std::vector<int> > ifaces_vrts(nbase_vrts + 2);
   ifaces_vrts[0].resize(nbase_vrts);
   std::iota(ifaces_vrts[0].begin(), ifaces_vrts[0].end(), 0);
